# Docs/refactor/artifacts/dashboard_openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "AxisEvaluateRequest": {
        "properties": {
          "game_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Game Id"
          },
          "history_b64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "History B64"
          },
          "parameter_overrides": {
            "additionalProperties": {
              "additionalProperties": {
                "type": "number"
              },
              "type": "object"
            },
            "title": "Parameter Overrides",
            "type": "object"
          },
          "parameters": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Parameters",
            "type": "object"
          },
          "position": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "prototype_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prototype Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "turn_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turn Index"
          }
        },
        "title": "AxisEvaluateRequest",
        "type": "object"
      },
      "AxisFixtureGenerateRequest": {
        "properties": {
          "count": {
            "default": 8,
            "title": "Count",
            "type": "integer"
          },
          "examples_per_move_count": {
            "default": 3,
            "title": "Examples Per Move Count",
            "type": "integer"
          },
          "max_depth": {
            "default": 1,
            "title": "Max Depth",
            "type": "integer"
          },
          "move_counts": {
            "items": {
              "type": "integer"
            },
            "title": "Move Counts",
            "type": "array"
          },
          "near_radius": {
            "default": 6,
            "title": "Near Radius",
            "type": "integer"
          },
          "noise_level": {
            "default": 0.08,
            "title": "Noise Level",
            "type": "number"
          },
          "opening_random_moves": {
            "default": 2,
            "title": "Opening Random Moves",
            "type": "integer"
          },
          "random_move_prob": {
            "default": 0.04,
            "title": "Random Move Prob",
            "type": "number"
          },
          "seed": {
            "default": 0,
            "title": "Seed",
            "type": "integer"
          },
          "time_ms": {
            "default": 2,
            "title": "Time Ms",
            "type": "integer"
          },
          "workers": {
            "default": 4,
            "title": "Workers",
            "type": "integer"
          }
        },
        "title": "AxisFixtureGenerateRequest",
        "type": "object"
      },
      "AxisPresetRequest": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "parameters": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Parameters",
            "type": "object"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "prototype_id": {
            "title": "Prototype Id",
            "type": "string"
          }
        },
        "required": [
          "name",
          "prototype_id"
        ],
        "title": "AxisPresetRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Hexo-RL Dashboard",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/axis/evaluate": {
      "post": {
        "operationId": "axis_evaluate_api_axis_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AxisEvaluateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Axis Evaluate Api Axis Evaluate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Axis Evaluate"
      }
    },
    "/api/axis/fixtures": {
      "get": {
        "operationId": "axis_fixtures_api_axis_fixtures_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Axis Fixtures Api Axis Fixtures Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Axis Fixtures"
      }
    },
    "/api/axis/fixtures/generate": {
      "post": {
        "operationId": "axis_fixtures_generate_api_axis_fixtures_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AxisFixtureGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Axis Fixtures Generate Api Axis Fixtures Generate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Axis Fixtures Generate"
      }
    },
    "/api/axis/presets": {
      "get": {
        "operationId": "axis_presets_api_axis_presets_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Axis Presets Api Axis Presets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Axis Presets"
      },
      "post": {
        "operationId": "axis_preset_api_axis_presets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AxisPresetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Axis Preset Api Axis Presets Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Axis Preset"
      }
    },
...
    "/api/games": {
      "get": {
        "operationId": "games_api_games_get",
        "parameters": [
          {
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Games Api Games Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Games"
      }
    },
    "/api/games/{game_id}/position/{turn_index}": {
      "get": {
        "operationId": "game_position_api_games__game_id__position__turn_index__get",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "turn_index",
            "required": true,
            "schema": {
              "title": "Turn Index",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Game Position Api Games  Game Id  Position  Turn Index  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Game Position"
      }
    },
    "/api/games/{game_id}/replay": {
      "get": {
        "operationId": "game_replay_api_games__game_id__replay_get",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Game Replay Api Games  Game Id  Replay Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Game Replay"
      }
    },
    "/api/games/{game_id}/snapshot.png": {
      "get": {
        "operationId": "game_snapshot_api_games__game_id__snapshot_png_get",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          },
          {
            "in": "query",
            "name": "turn_index",
            "required": false,
            "schema": {
              "default": -1,
              "title": "Turn Index",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "width",
            "required": false,
            "schema": {
              "default": 1280,
              "title": "Width",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "height",
            "required": false,
            "schema": {
              "default": 960,
              "title": "Height",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "context_rings",
            "required": false,
            "schema": {
              "default": 2,
              "title": "Context Rings",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "show_numbers",
            "required": false,
            "schema": {
              "default": true,
              "title": "Show Numbers",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "show_legal",
            "required": false,
            "schema": {
              "default": false,
              "title": "Show Legal",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "fit",
            "required": false,
            "schema": {
              "default": "played",
              "title": "Fit",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "near_radius",
            "required": false,
            "schema": {
              "default": 8,
              "title": "Near Radius",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Game Snapshot"
      }
    },
...
  }
}